Protect against closed streams
I noticed a pod process pegging a CPU. Can't reproduce it, but maybe it is just because it is gets stuck in an infinite error loop. If so, this may fix it. I'm guessing it is some kind of edge case when on a laptop and hibernating, waking up, etc.